+++ jpostman-annotations/src/main/java/io/jpostman/annotations/JPostmanCall.java
@@ -23,8 +23,22 @@ public @interface JPostmanCall {
 	/** @return request namespace */
 	String namespace() default "";
 
-	/** @return collection folder name */
-	String folder() default "";
+	/**
+	 * Collection folder path from the collection root to the target folder.
+	 *
+	 * <p>
+	 * Each value represents one folder level. An empty value selects requests from
+	 * the collection root.
+	 * </p>
+	 *
+	 * <pre>
+	 * folder = "Products"
+	 * folder = { "level1", "level2", "level3" }
+	 * </pre>
+	 *
+	 * @return folder path levels from parent to child
+	 */
+	String[] folder() default "";
 
 	/** @return collection request name */
 	String request() default "";